Regenerate code, add e2e test for failing scenario (#27)
Description of changes:

- Regenerate code after upstream code-generator changes
- Bump runtime version to 0.2.0
- Fix failing unit tests
- Improve unit test diff reporting for failing cases
- Add e2e test covering failing scenarios relevant to #22 

This PR will unblock #22. I was considering making the changes and adding them onto this PR but thought it might be more convenient to separate the many newly-generated lines from the actual change.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.