Fix delta comparison for auto failover, multi AZ, primary cluster ID (#69)
Nice.

Description of changes:
Properly retrieve the latest state for these 3 fields and compare them to the desired state. For each of these 3 spec fields, obtaining the corresponding latest state requires processing of a different `Status` field due to the structure of the `DescribeReplicationGroups` API response

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.