redis >=6 skip patch version compare (#115)
Issue #, if available:

Description of changes:

The purpose of this changes is to avoid the controller to compare patch versions when Redis version is >=6, since It's managed by AWS. More details [here](https://github.com/aws-controllers-k8s/community/issues/1737). 
 
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.