Support for CacheCluster resource (#137)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/2079

Description of changes:

This changelist adds support for the ElastiCache `CacheCluster` resource. The generated code did not account for many fields in the call to `ModifyCacheCluster`, and the `DescribeCacheClusters` API does not return all fields in the spec and values for some fields are reported in the `PendingModifiedValues` field instead. To handle this, the last requested configuration for `Spec.PreferredAvailabilityZones` is stored as annotations and used to compute `NewAvailabilityZones`. Moreover, `CacheNodeIdsToRemove` is computed using `Spec.NumCacheNodes` and `PendingModifiedValues.NumCacheNodes`. Support for `Spec.LogDeliveryConfigurations` is explicitly removed and parked for later because the Describe API does not return all fields for this struct. There were many changes to other ElastiCache resources introduced in the newer SDK, so they have been removed for now and not ignored. As a result, a subsequent `make build-controller` will result in new changes.

The logic for syncing tags and comparing engine versions has been refactored to promote reusability.

Closes https://github.com/aws-controllers-k8s/community/issues/2079.

Signed-off-by: cPu1 <patwal.chetan@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.